apps/requirements.txt
ccard==0.1.9
pika==1.3.1
psycopg2-binary==2.9.5
